refactor baseline
