+++ src/app/app.component.html
@@ -1,23 +1,21 @@
 <router-outlet>
-	<button (click)="fire.auth.signOut()">Logout</button>
+	<button (click)="data.auth.signOut()">Logout</button>
 	<p></p>
-	<button (click)="fire.auth.check()">check</button>
+	<button (click)="data.auth.check()">check</button>
 	<p></p>
-	<button (click)="fire.hash('Hello, World!')">hash</button>
+	<button (click)="data.hash('Hello, World!')">hash</button>
 	<p></p>
-	<button (click)="fire.snap('provider')">Snap Provider</button>
-	<div *ngIf="fire.snapshot.provider | async as p">
+	<button (click)="data.snap('provider')">Snap Provider</button>
+	<div *ngIf="data.snapshot.provider | async as p">
 		<ul>
 			<li *ngFor="let provider of p">{{provider.name}},{{provider._expire}}</li>
 		</ul>
 	</div>
 	<p></p>
-	<button (click)="fire.snap('class')">Snap Class</button>
-	<div *ngIf="fire.snapshot.class | async as c">
+	<button (click)="data.snap('class')">Snap Class</button>
+	<div *ngIf="data.snapshot.class | async as c">
 		<ul>
 			<li *ngFor="let class of c">{{class.name}}</li>
 		</ul>
 	</div>
-	<p></p>
-	<button (click)="member.attend('Step')">Attend</button>
 </router-outlet>
