+++ src/app/app.module.ts
@@ -25,8 +25,8 @@ import { environment } from '../environments/environment';
     BrowserModule,
     BrowserAnimationsModule,
     DBaseModule,
-    RoutingModule,
     AuthModule,
+    RoutingModule,
     ServiceWorkerModule.register('/ngsw-worker.js', { enabled: environment.production }),
   ],
   providers: [],
