+++ src/app/dbase/auth/auth.interface.ts
@@ -13,7 +13,7 @@ export interface ILink {
 	oauthAccessToken?: string;
 }
 
-export interface IUserInfo {					// borrowed from firebase-admin/auth
+export interface IUserInfo {			// borrowed from firebase-admin/auth
 	uid: string;
 	displayName: string;
 	email: string;
@@ -23,9 +23,11 @@ export interface IUserInfo {					// borrowed from firebase-admin/auth
 }
 
 export interface ICustomClaims {	// a special sub-set of fields from the User Token
-	memberName?: string;
-	memberUid?: string;
-	roles?: string[];
+	claims: {
+		memberName?: string;
+		memberUid?: string;
+		roles?: string[];
+	}
 }
 export interface IFireClaims {
 	user_id: string;
@@ -42,12 +44,12 @@ export interface IFireClaims {
 	auth_time: number;
 }
 export interface IJwtClaims {		// industry-standard json-web-token format
-	iss: string;						// issuer
-	iat?: number;						// issued time
-	exp: number;						// expiration time
-	nbf?: number;						// not before
-	aud?: string;						// audience
-	sub?: string;						// subject
+	exp: number;									// expiration time
+	iss: string;									// issuer
+	iat?: number;									// issued time
+	nbf?: number;									// not before
+	aud?: string;									// audience
+	sub?: string;									// subject
 }
 export type TTokenClaims = IFireClaims & ICustomClaims & IJwtClaims;
 export const enum JWT {
@@ -61,4 +63,5 @@ export const enum JWT {
 export const enum ROLES {
 	member = 'member',
 	admin = 'admin',
+	guest = 'guest',
 }
