+++ src/app/dbase/fire/fire.interface.ts
