+++ src/app/dbase/sync/sync.service.ts
@@ -10,10 +10,10 @@ import { SetAttend, DelAttend, TruncAttend } from '@dbase/state/store.define';
 
 import { IListen, StoreStorage } from '@dbase/sync/sync.define';
 import { LoginToken } from '@app/dbase/auth/auth.define';
-import { DBaseModule } from '@dbase/dbase.module';
-import { FireService } from '@dbase/data/fire.service';
 import { FIELD } from '@dbase/data/data.define';
-import { IQuery } from '@dbase/data/fire.interface';
+import { DBaseModule } from '@dbase/dbase.module';
+import { FireService } from '@dbase/fire/fire.service';
+import { IQuery } from '@dbase/fire/fire.interface';
 
 import { isFunction, sortKeys } from '@lib/object.library';
 import { createPromise } from '@lib/utility.library';
